pyproject.toml
[project]
name = "argus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "ultralytics>=8.4.17",
    "vision-agents[deepgram,elevenlabs,gemini,getstream,openrouter]>=0.3.8",
]
